+++ src/uqtestfuns/__init__.py
@@ -6,14 +6,8 @@ import sys
 
 from .core import Marginal
 from .core import ProbInput
-from .core import (
-    UQTestFunBareABC,
-    UQTestFunABC,
-    UQTestFunFixDimABC,
-    UQTestFunVarDimABC,
-)
 from .core import UQTestFun
-from .core import FunParams
+from .core import Parameters
 from .core import RealVariable
 from .core.registry import get_registry
 
@@ -23,10 +17,8 @@ from .test_functions import *  # noqa
 from .meta import UQMetaFunSpec
 from .meta import UQMetaTestFun
 
-from .helpers import list_functions
-
 from . import api
-from .api import create, list_parameters
+from .api import create, list_functions, list_parameters
 
 if sys.version_info >= (3, 8):
     from importlib import metadata
@@ -38,19 +30,15 @@ __version__ = metadata.version("uqtestfuns")
 __all__ = [
     "Marginal",
     "ProbInput",
-    "FunParams",
-    "UQTestFunBareABC",
-    "UQTestFunABC",
-    "UQTestFunFixDimABC",
-    "UQTestFunVarDimABC",
+    "Parameters",
     "UQTestFun",
     "RealVariable",
     "test_functions",
     "UQMetaFunSpec",
     "UQMetaTestFun",
-    "list_functions",
     "api",
     "create",
+    "list_functions",
     "list_parameters",
 ]