+++ src/uqtestfuns/core/__init__.py
@@ -3,26 +3,15 @@ The core subpackage of uqtestfuns.
 """
 
 from .design_vars.real_variable import RealVariable
-from .parameters import FunParams, Parameters
+from .parameters import Parameters
 from .prob_input.marginal import Marginal
 from .prob_input.probabilistic_input import ProbInput
-from .uqtestfun_abc import (
-    UQTestFunBareABC,
-    UQTestFunABC,
-    UQTestFunFixDimABC,
-    UQTestFunVarDimABC,
-)
 from .uqtestfun import UQTestFun
 
 __all__ = [
     "Marginal",
     "ProbInput",
-    "FunParams",
     "Parameters",
     "RealVariable",
-    "UQTestFunBareABC",
-    "UQTestFunABC",
-    "UQTestFunFixDimABC",
-    "UQTestFunVarDimABC",
     "UQTestFun",
 ]